+++ changelog.md
@@ -5,10 +5,11 @@
   * You can now sleep during day, waking up at dusk
 * `Food Heal Multiplier` can now be configured even if Combat Test preset it enabled
 * `Crops Require Water` no longer affects crops that aren't on farmland (e.g. Farmer's Delight rice)
+* `Combat.Shields` now disables itself when Shields+ is disabled
 
 ## 2.11.7
 * Data packs
-  * Integrated datapacks are temporarily removed until I figure out a way to make them controlled by config.  
+  * Integrated data packs are temporarily removed until I figure out a way to make them controlled by config.  
     Even though they're disabled I've made more changes:
   * Added Cheaper Chains
     * Chains are cheaper to craft (3 nuggets instead of 2 nuggets + 1 ingot)