+++ changelog.md
@@ -5,6 +5,7 @@
   * You can now sleep during day, waking up at dusk
 * `Food Heal Multiplier` can now be configured even if Combat Test preset is enabled
 * `Terrain Slowdown` now checks every block the player is on instead of just the block exactly below
+  * Takes all the blocks the player's on and makes an average of slowdowns
 * `Crops Require Water` no longer affects crops that aren't on farmland (e.g. Farmer's Delight rice)
 * `Combat.Shields` now disables itself when Shields+ is disabled